🔨 v0系のリソースを削除
