+++ tests/store/never.ts
@@ -1,5 +1,3 @@
-import {
-  DefineStoreModule
-} from '../..';
+import { DefineStoreModule } from '../../lib';
 
 export type Store = DefineStoreModule<'never', never, never, never, never>;
