--- v2-test/store/never.ts
@@ -1,3 +1,0 @@
-import { DefineStoreModule } from '../../lib';
-
-export type Store = DefineStoreModule<'never', never, never, never, never>;
